LPS-75742 Move to internal
